fix: [UTS-21] refactoring-files
