+++ lib/screens/login.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:runfinity/styles/app_colors.dart';
 import 'package:runfinity/widgets/appText.dart';
-import 'package:runfinity/widgets/formValidation.dart';
-import 'package:runfinity/widgets/loginCheckBox.dart';
-import 'package:runfinity/widgets/loginInput.dart';
-import 'package:runfinity/widgets/loginPasswordInput.dart';
-import 'package:runfinity/widgets/otherLoginOptions.dart';
+import 'package:runfinity/utils/formValidation.dart';
+import 'package:runfinity/widgets/login/loginCheckBox.dart';
+import 'package:runfinity/widgets/login/loginInput.dart';
+import 'package:runfinity/widgets/login/loginPasswordInput.dart';
+import 'package:runfinity/widgets/login/otherLoginOptions.dart';
 
 class Login extends StatefulWidget {
   const Login({super.key});
